refactor : 로또 랜덤 생성기 이름 수정
