+++ src/main/java/lotto/domain/LottoRandomNumber.java
@@ -1,13 +1,13 @@
-package lotto.utility;
+package lotto.domain;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
 import lotto.enums.LottoCriteria;
 
-public class LottoNumberUtility {
+public class LottoRandomNumber {
 
     public static List<Integer> lottoRandomNumber(){
         return Randoms.pickUniqueNumbersInRange(LottoCriteria.MIN_LOTTO_NUM.getCriteriaVal(),
-                LottoCriteria.MAX_LOTTO_COUNT.getCriteriaVal(), LottoCriteria.WINNING_NUMBER_COUNT.getCriteriaVal());
+                LottoCriteria.MAX_LOTTO_NUM.getCriteriaVal(), LottoCriteria.WINNING_NUMBER_COUNT.getCriteriaVal());
     }
 }
