+++ src/main/java/lotto/service/LottoService.java
@@ -2,13 +2,15 @@ package lotto.service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 import lotto.domain.Lotto;
 import lotto.domain.LottoCalculator;
+import lotto.domain.LottoFormatter;
 import lotto.domain.Lottos;
 import lotto.dto.LottoRequest;
 import lotto.dto.LottoWinResult;
 import lotto.enums.LottoCriteria;
-import lotto.utility.LottoNumberUtility;
+import lotto.domain.LottoRandomNumber;
 
 public class LottoService {
 
@@ -20,18 +22,16 @@ public class LottoService {
         return new Lottos(lottoList);
     }
 
-    public LottoWinResult calLottoResult(Lottos lottos, LottoRequest lottoRequest) {
+    public String calLottoResult(Lottos lottos, LottoRequest lottoRequest) {
         List<LottoWinResult> lottoWinResultList = lottos.getWinResult(lottoRequest.winningNumbers(), lottoRequest.bonusNumber());
         double lottoWinMoneyRate = LottoCalculator.getStatisticResult(lottoWinResultList,lottoRequest.buyMoney());
-        
+        Map<Integer, List<LottoWinResult>> winLottoResultMap = LottoCalculator.getWinLottoResultMap(lottoWinResultList);
+        return LottoFormatter.getLottoResultStr(lottoWinMoneyRate,winLottoResultMap);
     }
 
     private Lotto buyLotto(){
-        List<Integer> lottoRandomNumbers = LottoNumberUtility.lottoRandomNumber();
+        List<Integer> lottoRandomNumbers = LottoRandomNumber.lottoRandomNumber();
         Lotto lotto = new Lotto(lottoRandomNumbers);
         return lotto;
     }
-
 }
-
-
